+++ doc/releases.rst
@@ -9,6 +9,11 @@ Latest development release
     | Version: 1.32
     | Released: 2024-11-22
 
+- Added *defaults* argument to :class:`plural`.
+- Added *defaults* argument to :class:`truth`.
+- Added *all* argument to *stream_policy*.
+- Added :func:`tree()`  function.
+
 
 1.32 (2024-11-22)
 -----------------